==> penetration_resistance/00_master.R <==
### ERP penetrologger data
### Deployment: June 2024
### J Collins 
### 2024-06-22

## MASTER SCRIPT - RUN ALL ####

## RUN! ####

### 01 PACKAGES ####

source(file = "scripts/01_packages.R")

### 02 DATA ####

source(file = "scripts/02_data.R")

### 03 PLOTS ####

source(file = "scripts/03_plots.R")
